+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Populate substitution nodes in postprocess layer (DOCSPLAT-114).
+
 ## [v0.3.1] - 2020-02-14
 
 ### Added